requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-braces==1.14.0
django-cors-headers==3.4.0
django-extensions==2.2.9
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
idna==2.9
Markdown==3.2.2
networkx==2.4
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
reverse==0.1.0
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
splinter==0.13.0
sqlparse==0.3.1
urllib3==1.25.9
